requirements.txt
gradio
torch
transformers
pillow
numpy
opencv-python
sacremoses
